fix: move tiktoken wrapper to separate module for backend detection
Chonkie 1.4.0's AutoTokenizer detects backends by checking if backend names
appear in the type string (case-sensitive). By moving TiktokenWrapperForChonkie
to tiktoken_wrapper.py, the type string now contains "tiktoken" (lowercase):

  <class 'airweave.platform.chunkers.tiktoken_wrapper.TiktokenWrapperForChonkie'>

This ensures proper backend detection and that our encode() method with
allowed_special="all" is called correctly.
